Extract SchemeSelectable mixin
